hoan thien logic tra cuu tinh thu diem va logic tra ve ket qua xet tuyen chua css

+++ student-system/src/main/java/admissions/student_system/entity/Nganh.java
@@ -33,10 +33,10 @@ public class Nganh {
     private Integer nChiTieu;
 
     @Column(name = "n_diemsan")
-    private BigDecimal nDiemSan;
+    private Double nDiemSan;
 
     @Column(name = "n_diemtrungtuyen")
-    private BigDecimal nDiemTrungTuyen;
+    private Double nDiemTrungTuyen;
 
     @Column(name = "n_tuyenthang")
     private String nTuyenThang;
